+++ convert2rhel/grub.py
@@ -23,6 +23,7 @@ import re
 import shutil
 
 from convert2rhel import backup, systeminfo, utils
+from convert2rhel.backup.files import RestorableFile
 
 
 logger = logging.getLogger(__name__)
@@ -618,8 +619,8 @@ def update_grub_after_conversion():
     rather re-generate the GRUB2 config file and install the GRUB2 image.
     """
 
-    backup.backup_control.push(backup.RestorableFile(GRUB2_BIOS_CONFIG_FILE))
-    backup.backup_control.push(backup.RestorableFile(GRUB2_BIOS_ENV_FILE))
+    backup.backup_control.push(RestorableFile(GRUB2_BIOS_CONFIG_FILE))
+    backup.backup_control.push(RestorableFile(GRUB2_BIOS_ENV_FILE))
 
     grub2_config_file = get_grub_config_file()
